subproject isolation for building better support for multiple bindings
