+++ .gitattributes
@@ -1,1 +1,7 @@
 client/docs/swagger-ui/* linguist-vendored
+#
+# https://help.github.com/articles/dealing-with-line-endings/
+#
+# These are explicitly windows files and should use crlf
+*.bat           text eol=crlf
+
