--- protoBindings/bindings/rust/build.gradle.kts
@@ -1,6 +1,0 @@
-
-// Register task
-tasks.register<rust.ProtobufRustGrpcTask>("generateRustGrpc")
-
-// Remove Maven publishing
-tasks.withType<PublishToMavenRepository>().configureEach { setProperty("enabled", false) }
