+++ protoBindings/gradle.properties
@@ -2,6 +2,38 @@ kotlin.code.style=official
 kapt.use.worker.api=false
 kapt.incremental.apt=false
 org.gradle.jvmargs=-Xmx4096M
+
 # Source directory (relative or absolute path) of the proto files that will be compiled.
 # A comma separated list of directories of the proto files for compilation.
 protoDirs=../proto/,../third_party/proto/
+
+####################
+# Maven publishing #
+####################
+# Group ID form Maven publishing
+group.id=io.provenance
+
+pom.name=Provenance Blockchain Proto Bindings
+pom.description=Protobuf bindings
+pom.url=https://provenance.io
+
+license.name=The Apache License, Version 2.0
+license.url=https://www.apache.org/licenses/LICENSE-2.0.txt
+
+# Developers
+developer.id=egaxhaj-figure
+developer.name=Ergels Gaxhaj
+developer.email=egaxhaj@figure.com
+
+scm.connection=git@github.com:provenance-io/provenance.git
+scm.developerConnection=git@github.com/provenance-io/provenance.git
+scm.url=https://github.com/provenance-io/provenance
+
+####################
+# Nexus Publishing #
+####################
+nexus.url=https://s01.oss.sonatype.org/service/local/
+nexus.snapshot.repository.url=https://s01.oss.sonatype.org/content/repositories/snapshots/
+nexus.username=ossrhUsername
+nexus.password=ossrhPassword
+nexus.staging.profile.id=3180ca260b82a7
